requirements.txt
streamlit
pandas
pdfplumber
PyGithub
